pyproject.toml
[tool.poetry]
name = "sample"
version = "1.0.0"
description = "The back end for the sample Application"
authors = ["Bhargav Singapuri <bhsingapuri@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
pydantic = { extras = ["dotenv"], version = "^1.8.2" }
psycopg2-binary = "^2.9.1"
django-allauth = "^0.45.0"
gunicorn = "^20.1.0"
whitenoise = "^5.3.0"
django-debug-toolbar = "^3.2.2"
django-ninja = "^0.16.0"
dj-rest-auth = "^2.2.1"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.10.1"
dj-database-url = "^0.5.0"

[tool.poetry.dev-dependencies]
factory-boy = "^3.2.0"
pre-commit = "^2.14.0"
pytest = "^6.2.4"
pytest-cov = "^2.12.1"
pytest-django = "^4.4.0"
tqdm = "^4.62.1"
pandas = "^1.3.2"
xlrd = "^2.0.1"
jupyter = "^1.0.0"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
